requirements.txt
anyio==4.2.0
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.3
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.1
rsa==4.9
sniffio==1.3.0
Telethon==1.33.1
typing_extensions==4.9.0
Werkzeug==3.0.3
